Add generic driver for Silead tochscreens
This driver adds support for Silead touchscreens. It has been tested
with GSL1680 and GSL3680 touch panels.

It supports ACPI and device tree enumeration. Screen resolution,
the maximum number of fingers supported and firmware name are
configurable using ACPI/DT properties.

If the device properties are not present it falls back to using defaults:
 - x 4095
 - y 4095
 - max fingers 10
 - firmware name [HID/name].fw

If there is no named GPIO for power it falls back to using an indexed GPIO
and it requests the GPIO pin with index 1. If there isn't one it disables
PM support.

All the hardware variants tested report finger id 0 for all fingers so
the finger tracking is done using the input subsystem's slot assignment.

Signed-off-by: Robert Dolca <robert.dolca@intel.com>